+++ predict.c
@@ -14,7 +14,7 @@
 void reconstruct_timebins(void)
 {
     int i, n, prev, bin;
-    long long glob_sum1, glob_sum2;
+    int64_t glob_sum1, glob_sum2;
 
     for(bin = 0; bin < TIMEBINS; bin++)
     {
